+++ src/resources/views/channel/show.blade.php
@@ -10,6 +10,12 @@
 
 @section('content')
     <x-appshell::card>
-        <x-slot:title>{{ __('Statistics') }}</x-slot:title>
+        <x-slot:title>{{ __('Assigned Products') }}</x-slot:title>
+        <x-slot:actions>
+            <x-appshell::button href="#" variant="outline-success" icon="+" size="sm">{{ __('Assign a product') }}</x-appshell::button>
+        </x-slot:actions>
+
+        {!! widget('vanilo::channel.products')->render($channel->products) !!}
+
     </x-appshell::card>
 @stop