+++ src/resources/views/master-product/show.blade.php
@@ -18,6 +18,27 @@
                 @endif
                 <x-slot:subtitle>
                     {{ $product->variants->implode('sku', ', ') }}&nbsp;
+                    @if($multiChannelEnabled)
+                        |
+                        @if($product->channels->count())
+                            {{ __('Channels') }}:
+                            @can('view channels')
+                                @foreach($product->channels->take(3) as $channel)
+                                    <a href="{{ route('vanilo.admin.channel.show', $channel->id) }}">
+                                        {{ $channel->name }}
+                                    </a>
+                                @endforeach
+                            @else
+                                {{ $product->channels->take(3)->implode('name', ' | ') }}
+                            @endcan
+                        @else
+                            {{ __('no channels') }}
+                        @endif
+
+                        @if($product->channels->count() > 3)
+                            | {{ __('+ :num more...', ['num' => $product->channels->count() - 3]) }}
+                        @endif
+                    @endif
                 </x-slot:subtitle>
             </x-appshell::card-with-icon>
         </div>