+++ src/resources/views/product/create.blade.php
@@ -22,6 +22,7 @@
 
         <div class="col-12 col-lg-4 col-xl-3">
             @include('vanilo::media._create')
+            @include('vanilo::channel._create')
         </div>
 
     {!! Form::close() !!}